requirements.txt
# Core FastAPI & Server
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Pydantic (use version with pre-built wheels)
pydantic==2.6.4
pydantic-core==2.16.3
pydantic-settings==2.2.1
email-validator==2.1.0

# LangChain & AI
langchain==0.1.20
langchain-core==0.1.52
langchain-google-genai==0.0.11
langgraph==0.0.20
google-generativeai==0.4.1

# Database
supabase==2.4.4
httpcore==1.0.2

# Utilities
python-dotenv==1.0.0
websockets==12.0
httpx==0.26.0
typing-extensions==4.9.0
annotated-types==0.6.0
